--- vroom-core/src/main/java/com/fizzbuzz/vroom/core/persist/datastore/OfyUtils.java
@@ -1,33 +1,0 @@
-package com.fizzbuzz.vroom.core.persist.datastore;
-
-/*
- * Copyright (c) 2013 Fizz Buzz LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import com.fizzbuzz.vroom.core.persist.datastore.dao.BaseDao;
-import com.googlecode.objectify.Ref;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class OfyUtils {
-    public static <DAO extends BaseDao> List<Long> getIdsFromRefs(List<Ref<DAO>> refs) {
-        List<Long> ids = new ArrayList<>();
-        for (Ref<DAO> ref : refs) {
-            ids.add(ref.getKey().getId());
-        }
-        return ids;
-    }
-
-
-}
