--- vroom-core/src/main/java/com/fizzbuzz/vroom/core/persist/datastore/entity/PackageLogger.java
@@ -1,23 +1,0 @@
-package com.fizzbuzz.vroom.core.persist.datastore.entity;
-
-/*
- * Copyright (c) 2013 Fizz Buzz LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-class PackageLogger {
-    // This class exists just to define a static string for use by other classes in this package. Make the constructor
-    // private to prevent instantiation.
-    private PackageLogger() {
-    }
-
-    public static final String TAG = "com.fizzbuzz.vroom.core.persist.datastore.entity";}
