+++ vroom-core/src/main/java/com/fizzbuzz/vroom/core/service/datastore/OfyUtils.java
@@ -1,4 +1,4 @@
-package com.fizzbuzz.vroom.core.biz;
+package com.fizzbuzz.vroom.core.service.datastore;
 
 /*
  * Copyright (c) 2013 Fizz Buzz LLC
@@ -14,11 +14,20 @@ package com.fizzbuzz.vroom.core.biz;
  * limitations under the License.
  */
 
-import com.fizzbuzz.vroom.core.domain.KeyedObject;
-import com.fizzbuzz.vroom.core.service.datastore.entity.EntityCollection;
+import com.fizzbuzz.vroom.core.service.datastore.dao.BaseDao;
+import com.googlecode.objectify.Ref;
 
-public abstract class ParentedEntityCollectionBiz<KO extends KeyedObject> extends EntityCollectionBiz<KO> {
-    public ParentedEntityCollectionBiz(EntityCollection<KO> entityCollection) {
-        super(entityCollection);
+import java.util.ArrayList;
+import java.util.List;
+
+public class OfyUtils {
+    public static <DAO extends BaseDao> List<Long> getIdsFromRefs(List<Ref<DAO>> refs) {
+        List<Long> ids = new ArrayList<>();
+        for (Ref<DAO> ref : refs) {
+            ids.add(ref.getKey().getId());
+        }
+        return ids;
     }
+
+
 }
