+++ vroom-core/src/main/java/com/fizzbuzz/vroom/core/service/datastore/dao/TimeStampedDao.java
@@ -1,0 +1,49 @@
+package com.fizzbuzz.vroom.core.service.datastore.dao;
+
+/*
+ * Copyright (c) 2013 Fizz Buzz LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import com.fizzbuzz.vroom.core.domain.KeyedObject;
+
+import java.util.Date;
+
+public abstract class TimeStampedDao<KO extends KeyedObject> extends BaseDao<KO> {
+    private Date createdDate;
+    private Date modifiedDate;
+
+    // no-arg constructor needed by Objectify
+    public TimeStampedDao() {
+    }
+
+    protected TimeStampedDao(final Long id) {
+        super(id);
+    }
+
+    public Date getCreatedDate() {
+        return createdDate;
+    }
+
+    public void setCreatedDate(final Date createdDate) {
+
+        this.createdDate = createdDate;
+    }
+
+    public Date getModifiedDate() {
+        return modifiedDate;
+    }
+
+    public void setModifiedDate(final Date modifiedDate) {
+        this.modifiedDate = modifiedDate;
+    }
+}
