--- vroom-core/src/test/java/com/fizzbuzz/vroom/core/persist/datastore/DataStoreRule.java
@@ -1,39 +1,0 @@
-package com.fizzbuzz.vroom.core.persist.datastore;
-
-/*
- * Copyright (c) 2013 Fizz Buzz LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import com.google.appengine.tools.development.testing.LocalBlobstoreServiceTestConfig;
-import com.google.appengine.tools.development.testing.LocalDatastoreServiceTestConfig;
-import com.google.appengine.tools.development.testing.LocalMemcacheServiceTestConfig;
-import com.google.appengine.tools.development.testing.LocalServiceTestHelper;
-import com.google.appengine.tools.development.testing.LocalTaskQueueTestConfig;
-import org.junit.rules.ExternalResource;
-
-public class DataStoreRule extends ExternalResource {
-    private LocalServiceTestHelper mHelper;
-
-    @Override
-    protected void before() throws Throwable {
-        mHelper = new LocalServiceTestHelper(new LocalDatastoreServiceTestConfig(),
-                new LocalBlobstoreServiceTestConfig(), new LocalTaskQueueTestConfig(),
-                new LocalMemcacheServiceTestConfig());
-        mHelper.setUp();
-    }
-
-    @Override
-    protected void after() {
-        mHelper.tearDown();
-    }
-}
