--- vroom-core/src/test/java/com/fizzbuzz/vroom/core/persist/datastore/TestKeyedObject.java
@@ -1,14 +1,0 @@
-package com.fizzbuzz.vroom.core.persist.datastore;
-
-import com.fizzbuzz.vroom.core.domain.KeyedObject;
-import com.fizzbuzz.vroom.core.domain.LongKey;
-
-/**
- * Copyright (c) 2013 Fitivity Inc.
- */
-
-public class TestKeyedObject extends KeyedObject<LongKey> {
-    public TestKeyedObject() {
-        super(null);
-    }
-}
