+++ vroom-core/src/test/java/com/fizzbuzz/vroom/core/service/datastore/TestKeyedObject.java
@@ -1,0 +1,14 @@
+package com.fizzbuzz.vroom.core.service.datastore;
+
+import com.fizzbuzz.vroom.core.domain.KeyedObject;
+import com.fizzbuzz.vroom.core.domain.LongKey;
+
+/**
+ * Copyright (c) 2013 Fitivity Inc.
+ */
+
+public class TestKeyedObject extends KeyedObject<LongKey> {
+    public TestKeyedObject() {
+        super(null);
+    }
+}
